Server/requirements.txt
Flask==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-Script==2.0.5
gunicorn==19.7.1
Jinja2==2.9.6
jsonpickle==0.9.5
redis==2.10.5
requests==2.18.1
python-dateutil==2.6.1
